pyproject.toml
[project]
name = "noorinalabs-isnad-ingest-platform"
version = "0.1.0"
description = "Standalone data ingestion platform for the isnad-graph project"
license = {text = "Apache-2.0"}
requires-python = ">=3.14"
dependencies = [
    "pydantic>=2.6",
    "pydantic-settings>=2.1",
    "neo4j>=5.15",
    "psycopg[binary]>=3.1",
    "httpx>=0.27",
    "pyarrow>=15.0",
    "pandas>=2.2",
    "pyyaml>=6.0",
    "structlog>=24.1",
    "python-dotenv>=1.0",
    "kaggle>=1.6",
    "beautifulsoup4>=4.12",
    "lxml>=5.1",
    "tqdm>=4.66",
    "tenacity>=8.2",
    "rapidfuzz>=3.6",
    "kafka-python>=2.0",
    "boto3>=1.34",
]

[project.scripts]
isnad-ingest = "src.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
    "types-tqdm>=4.66",
    "types-pyyaml>=6.0.12.20250915",
    "testcontainers[neo4j,postgres]>=4.0",
    "hypothesis>=6.0",
    "pre-commit>=4.0",
    "pip-audit>=2.7",
]
ml = [
    "sentence-transformers>=2.5",
    "faiss-cpu>=1.7",
    "transformers>=4.38",
    "torch>=2.2",
    "camel-tools>=1.5",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py313"  # ruff doesn't support py314 yet; py313 is the closest

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "PD"]

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["pyarrow", "pyarrow.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["transformers", "transformers.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["faiss", "faiss.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["sentence_transformers", "sentence_transformers.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["testcontainers", "testcontainers.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["camel_tools", "camel_tools.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["bs4", "bs4.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tenacity", "tenacity.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tqdm", "tqdm.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["kaggle", "kaggle.*"]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_uninitialized = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "ml: marks tests requiring ML dependencies (sentence-transformers, faiss)",
    "integration: marks tests requiring Docker containers (Neo4j, PostgreSQL)",
]
